[refactor] Separate CPU plugin
